NSX Kune: Have V and T Coexist in the same plugin
Kune - together in esperanto

+++ vmware_nsx/plugins/nsx_v/plugin.py
@@ -219,6 +219,8 @@ class NsxVPluginV2(addr_pair_db.AllowedAddressPairsMixin,
     def __init__(self):
         self._extension_manager = nsx_managers.ExtensionManager()
         super(NsxVPluginV2, self).__init__()
+        # DEBUG ADIT - how can we know this?
+        self._is_sub_plugin = True
         # Bind the dummy L3 notifications
         self.l3_rpc_notifier = l3_rpc_agent_api.L3NotifyAPI()
         self.init_is_complete = False
@@ -994,7 +996,9 @@ class NsxVPluginV2(addr_pair_db.AllowedAddressPairsMixin,
         return '%s.%03d' % (device_id, port_index)
 
     def init_availability_zones(self):
-        self._availability_zones_data = nsx_az.NsxVAvailabilityZones()
+        validate_default = not self._is_sub_plugin
+        self._availability_zones_data = nsx_az.NsxVAvailabilityZones(
+            validate_default=validate_default)
 
     def _list_availability_zones(self, context, filters=None):
         #TODO(asarfaty): We may need to use the filters arg, but now it
@@ -4130,7 +4134,7 @@ class NsxVPluginV2(addr_pair_db.AllowedAddressPairsMixin,
 
         return sg_data
 
-    def delete_security_group(self, context, id):
+    def delete_security_group(self, context, id, delete_base=True):
         """Delete a security group."""
         self._prevent_non_admin_delete_provider_sg(context, id)
         self._prevent_non_admin_delete_policy_sg(context, id)
@@ -4142,8 +4146,9 @@ class NsxVPluginV2(addr_pair_db.AllowedAddressPairsMixin,
             # Find nsx security group
             nsx_sg_id = nsx_db.get_nsx_security_group_id(context.session, id)
 
-            # Delete neutron security group
-            super(NsxVPluginV2, self).delete_security_group(context, id)
+            if delete_base:
+                # Delete neutron security group
+                super(NsxVPluginV2, self).delete_security_group(context, id)
 
             # Delete nsx rule sections
             self._delete_section(section_uri)
